Refactors editor to use the MVP pattern (#67)
