+++ static/styleEditor.css
@@ -1,0 +1,18 @@
+* {
+    font-family:"Helvetica Neue", Helvetica, Arial;
+    font-size: 12px;
+    color: #333;
+}
+
+QDialog {
+    border: 1px solid #ddd;
+}
+
+
+QLineEdit, QComboBox, QSpinBox, QDoubleSpinBox {
+    padding: 5px;
+}
+
+QPushButton {
+    padding: 10px 15px;
+}
